Fix RewriteAsHoppingWindow optimization
